pyproject.toml
[project]
name = "bg-remover"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pillow",
  "rembg",          # CPU build; see GPU note below
  "python-multipart",
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
